+++ Comet/AdminDispatcherOffice365ListVirtualAccountsRequest.php
@@ -1,0 +1,124 @@
+<?php
+
+/**
+ * Copyright (c) 2018-2021 Comet Licensing Ltd.
+ * Please see the LICENSE file for usage information.
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+namespace Comet;
+
+/**
+ * Comet Server AdminDispatcherOffice365ListVirtualAccounts API
+ * Request a list of Office365 Resources (groups, sites, teams groups and users)
+ * The remote device must have given consent for an MSP to browse their files.
+ *
+ * You must supply administrator authentication credentials to use this API.
+ * This API requires the Auth Role to be enabled.
+ */
+class AdminDispatcherOffice365ListVirtualAccountsRequest implements \Comet\NetworkRequest {
+
+	/**
+	 * The live connection GUID
+	 *
+	 * @var string
+	 */
+	protected $TargetID = null;
+
+	/**
+	 * The Office365 account credential
+	 *
+	 * @var \Comet\Office365Credential
+	 */
+	protected $Credentials = null;
+
+	/**
+	 * Construct a new AdminDispatcherOffice365ListVirtualAccountsRequest instance.
+	 *
+	 * @param string $TargetID The live connection GUID
+	 * @param \Comet\Office365Credential $Credentials The Office365 account credential
+	 */
+	public function __construct($TargetID, Office365Credential $Credentials)
+	{
+		$this->TargetID = $TargetID;
+		$this->Credentials = $Credentials;
+	}
+
+	/**
+	 * Get the URL where this POST request should be submitted to.
+	 *
+	 * @return string
+	 */
+	public function Endpoint()
+	{
+		return '/api/v1/admin/dispatcher/office365-list-virtual-accounts';
+	}
+
+	public function Method()
+	{
+		return 'POST';
+	}
+
+	public function ContentType()
+	{
+		return 'application/x-www-form-urlencoded';
+	}
+
+	/**
+	 * Get the POST parameters for this request.
+	 *
+	 * @return string[]
+	 */
+	public function Parameters()
+	{
+		$ret = [];
+		$ret["TargetID"] = (string)($this->TargetID);
+		$ret["Credentials"] = $this->Credentials->toJSON();
+		return $ret;
+	}
+
+	/**
+	 * Decode types used in a response to this request.
+	 * Use any network library to make the request.
+	 *
+	 * @param int $responseCode HTTP response code
+	 * @param string $body HTTP response body
+	 * @return \Comet\BrowseOffice365ListVirtualAccountsResponse
+	 * @throws \Exception
+	 */
+	public static function ProcessResponse($responseCode, $body)
+	{
+		// Require expected HTTP 200 response
+		if ($responseCode !== 200) {
+			throw new \Exception("Unexpected HTTP " . intval($responseCode) . " response");
+		}
+
+		// Decode JSON
+		$decoded = \json_decode($body); // as stdClass
+		if (\json_last_error() != \JSON_ERROR_NONE) {
+			throw new \Exception("JSON decode failed: " . \json_last_error_msg());
+		}
+
+		// Try to parse as error format
+		$isCARMDerivedType = (($decoded instanceof \stdClass) && property_exists($decoded, 'Status') && property_exists($decoded, 'Message'));
+		if ($isCARMDerivedType) {
+			$carm = \Comet\APIResponseMessage::createFromStdclass($decoded);
+			if ($carm->Status >= 400) {
+				throw new \Exception("Error " . $carm->Status . ": " . $carm->Message);
+			}
+		}
+
+		// Parse as BrowseOffice365ListVirtualAccountsResponse
+		if (is_array($decoded) && count($decoded) === 0) {
+		// Work around edge case in json_decode--json_encode stdClass conversion
+			$ret = \Comet\BrowseOffice365ListVirtualAccountsResponse::createFromStdclass(new \stdClass());
+		} else {
+			$ret = \Comet\BrowseOffice365ListVirtualAccountsResponse::createFromStdclass($decoded);
+		}
+
+		return $ret;
+	}
+
+}
+