+++ Comet/BrowseOffice365ListVirtualAccountsResponse.php
@@ -9,68 +9,33 @@
 
 namespace Comet;
 
-class ServerMetaBrandingProperties {
-
-	/**
-	 * @var string
-	 */
-	public $BrandName = "";
-
-	/**
-	 * @var string
-	 */
-	public $ProductName = "";
-
-	/**
-	 * @var boolean
-	 */
-	public $HasImage = false;
-
-	/**
-	 * @var string
-	 */
-	public $ImageEtag = "";
-
-	/**
-	 * @var string
-	 */
-	public $TopColor = "";
-
-	/**
-	 * @var boolean
-	 */
-	public $HideNewsArea = false;
-
-	/**
-	 * @var boolean
-	 */
-	public $AllowUnauthenticatedDownloads = false;
-
-	/**
-	 * @var boolean
-	 */
-	public $AllowAuthenticatedDownloads = false;
+class BrowseOffice365ListVirtualAccountsResponse {
 
 	/**
 	 * @var int
 	 */
-	public $PruneLogsAfterDays = 0;
+	public $Status = 0;
 
 	/**
-	 * @var int
+	 * @var string
 	 */
-	public $ExpiredInSeconds = 0;
+	public $Message = "";
+
+	/**
+	 * @var \Comet\Office365MixedVirtualAccount[]
+	 */
+	public $Objects = [];
 
 	/**
 	 * Preserve unknown properties when dealing with future server versions.
 	 *
-	 * @see ServerMetaBrandingProperties::RemoveUnknownProperties() Remove all unknown properties
+	 * @see BrowseOffice365ListVirtualAccountsResponse::RemoveUnknownProperties() Remove all unknown properties
 	 * @var array
 	 */
 	private $__unknown_properties = [];
 
 	/**
-	 * Replace the content of this ServerMetaBrandingProperties object from a PHP \stdClass.
+	 * Replace the content of this BrowseOffice365ListVirtualAccountsResponse object from a PHP \stdClass.
 	 * The data could be supplied from an API call after json_decode(...); or generated manually.
 	 *
 	 * @param \stdClass $sc Object data as stdClass
@@ -78,48 +43,31 @@ class ServerMetaBrandingProperties {
 	 */
 	protected function inflateFrom(\stdClass $sc)
 	{
-		if (property_exists($sc, 'BrandName')) {
-			$this->BrandName = (string)($sc->BrandName);
+		if (property_exists($sc, 'Status')) {
+			$this->Status = (int)($sc->Status);
 		}
-		if (property_exists($sc, 'ProductName')) {
-			$this->ProductName = (string)($sc->ProductName);
+		if (property_exists($sc, 'Message')) {
+			$this->Message = (string)($sc->Message);
 		}
-		if (property_exists($sc, 'HasImage')) {
-			$this->HasImage = (bool)($sc->HasImage);
-		}
-		if (property_exists($sc, 'ImageEtag')) {
-			$this->ImageEtag = (string)($sc->ImageEtag);
-		}
-		if (property_exists($sc, 'TopColor')) {
-			$this->TopColor = (string)($sc->TopColor);
-		}
-		if (property_exists($sc, 'HideNewsArea')) {
-			$this->HideNewsArea = (bool)($sc->HideNewsArea);
-		}
-		if (property_exists($sc, 'AllowUnauthenticatedDownloads')) {
-			$this->AllowUnauthenticatedDownloads = (bool)($sc->AllowUnauthenticatedDownloads);
-		}
-		if (property_exists($sc, 'AllowAuthenticatedDownloads')) {
-			$this->AllowAuthenticatedDownloads = (bool)($sc->AllowAuthenticatedDownloads);
-		}
-		if (property_exists($sc, 'PruneLogsAfterDays')) {
-			$this->PruneLogsAfterDays = (int)($sc->PruneLogsAfterDays);
-		}
-		if (property_exists($sc, 'ExpiredInSeconds')) {
-			$this->ExpiredInSeconds = (int)($sc->ExpiredInSeconds);
+		if (property_exists($sc, 'Objects')) {
+			$val_2 = [];
+			if ($sc->Objects !== null) {
+				for($i_2 = 0; $i_2 < count($sc->Objects); ++$i_2) {
+					if (is_array($sc->Objects[$i_2]) && count($sc->Objects[$i_2]) === 0) {
+					// Work around edge case in json_decode--json_encode stdClass conversion
+						$val_2[] = \Comet\Office365MixedVirtualAccount::createFromStdclass(new \stdClass());
+					} else {
+						$val_2[] = \Comet\Office365MixedVirtualAccount::createFromStdclass($sc->Objects[$i_2]);
+					}
+				}
+			}
+			$this->Objects = $val_2;
 		}
 		foreach(get_object_vars($sc) as $k => $v) {
 			switch($k) {
-			case 'BrandName':
-			case 'ProductName':
-			case 'HasImage':
-			case 'ImageEtag':
-			case 'TopColor':
-			case 'HideNewsArea':
-			case 'AllowUnauthenticatedDownloads':
-			case 'AllowAuthenticatedDownloads':
-			case 'PruneLogsAfterDays':
-			case 'ExpiredInSeconds':
+			case 'Status':
+			case 'Message':
+			case 'Objects':
 				break;
 			default:
 				$this->__unknown_properties[$k] = $v;
@@ -128,25 +76,25 @@ class ServerMetaBrandingProperties {
 	}
 
 	/**
-	 * Coerce a stdClass into a new strongly-typed ServerMetaBrandingProperties object.
+	 * Coerce a stdClass into a new strongly-typed BrowseOffice365ListVirtualAccountsResponse object.
 	 *
 	 * @param \stdClass $sc Object data as stdClass
-	 * @return ServerMetaBrandingProperties
+	 * @return BrowseOffice365ListVirtualAccountsResponse
 	 */
 	public static function createFromStdclass(\stdClass $sc)
 	{
-		$retn = new ServerMetaBrandingProperties();
+		$retn = new BrowseOffice365ListVirtualAccountsResponse();
 		$retn->inflateFrom($sc);
 		return $retn;
 	}
 
 	/**
-	 * Coerce a plain PHP array into a new strongly-typed ServerMetaBrandingProperties object.
+	 * Coerce a plain PHP array into a new strongly-typed BrowseOffice365ListVirtualAccountsResponse object.
 	 * Because the Comet Server requires strict distinction between empty objects ({}) and arrays ([]),
 	 * the result of this method may not be safe to re-submit to the Comet Server.
 	 *
 	 * @param array $arr Object data as PHP array
-	 * @return ServerMetaBrandingProperties
+	 * @return BrowseOffice365ListVirtualAccountsResponse
 	 */
 	public static function createFromArray(array $arr)
 	{
@@ -158,7 +106,7 @@ class ServerMetaBrandingProperties {
 	}
 
 	/**
-	 * Coerce a plain PHP array into a new strongly-typed ServerMetaBrandingProperties object.
+	 * Coerce a plain PHP array into a new strongly-typed BrowseOffice365ListVirtualAccountsResponse object.
 	 * Because the Comet Server requires strict distinction between empty objects ({}) and arrays ([]),
 	 * the result of this method may not be safe to re-submit to the Comet Server.
 	 *
@@ -166,7 +114,7 @@ class ServerMetaBrandingProperties {
 	 *             (A) acknowledge this and continue by switching to createFromArray, or
 	 *             (b) switch to the roundtrip-safe createFromStdclass alternative.
 	 * @param array $arr Object data as PHP array
-	 * @return ServerMetaBrandingProperties
+	 * @return BrowseOffice365ListVirtualAccountsResponse
 	 */
 	public static function createFrom(array $arr)
 	{
@@ -174,10 +122,10 @@ class ServerMetaBrandingProperties {
 	}
 
 	/**
-	 * Coerce a JSON string into a new strongly-typed ServerMetaBrandingProperties object.
+	 * Coerce a JSON string into a new strongly-typed BrowseOffice365ListVirtualAccountsResponse object.
 	 *
 	 * @param string $JsonString Object data as JSON string
-	 * @return ServerMetaBrandingProperties
+	 * @return BrowseOffice365ListVirtualAccountsResponse
 	 */
 	public static function createFromJSON($JsonString)
 	{
@@ -185,13 +133,13 @@ class ServerMetaBrandingProperties {
 		if (\json_last_error() != \JSON_ERROR_NONE) {
 			throw new \Exception("JSON decode failed: " . \json_last_error_msg());
 		}
-		$retn = new ServerMetaBrandingProperties();
+		$retn = new BrowseOffice365ListVirtualAccountsResponse();
 		$retn->inflateFrom($decodedJsonObject);
 		return $retn;
 	}
 
 	/**
-	 * Convert this ServerMetaBrandingProperties object into a plain PHP array.
+	 * Convert this BrowseOffice365ListVirtualAccountsResponse object into a plain PHP array.
 	 *
 	 * Unknown properties may still be represented as \stdClass objects.
 	 *
@@ -201,16 +149,20 @@ class ServerMetaBrandingProperties {
 	public function toArray($for_json_encode = false)
 	{
 		$ret = [];
-		$ret["BrandName"] = $this->BrandName;
-		$ret["ProductName"] = $this->ProductName;
-		$ret["HasImage"] = $this->HasImage;
-		$ret["ImageEtag"] = $this->ImageEtag;
-		$ret["TopColor"] = $this->TopColor;
-		$ret["HideNewsArea"] = $this->HideNewsArea;
-		$ret["AllowUnauthenticatedDownloads"] = $this->AllowUnauthenticatedDownloads;
-		$ret["AllowAuthenticatedDownloads"] = $this->AllowAuthenticatedDownloads;
-		$ret["PruneLogsAfterDays"] = $this->PruneLogsAfterDays;
-		$ret["ExpiredInSeconds"] = $this->ExpiredInSeconds;
+		$ret["Status"] = $this->Status;
+		$ret["Message"] = $this->Message;
+		{
+			$c0 = [];
+			for($i0 = 0; $i0 < count($this->Objects); ++$i0) {
+				if ( $this->Objects[$i0] === null ) {
+					$val0 = $for_json_encode ? (object)[] : [];
+				} else {
+					$val0 = $this->Objects[$i0]->toArray($for_json_encode);
+				}
+				$c0[] = $val0;
+			}
+			$ret["Objects"] = $c0;
+		}
 
 		// Reinstate unknown properties from future server versions
 		foreach($this->__unknown_properties as $k => $v) {